move the poker logic to `Room` page
