--- src/pages/NotFound/NotFound.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-class NotFound extends Component {
-  render() {
-    return <div>404</div>;
-  }
-}
-
-export default NotFound;
