+++ src/routes.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { Router, Route } from 'react-router';
 
 import App from 'pages/App/App';
-import NotFound from 'pages/NotFound/NotFound';
+import Room from 'pages/Room/Room';
 
 const Routes = (props) => (
   <Router {...props}>
     <Route path="/" component={App} />
-    <Route path="*" component={NotFound} />
+    <Route path="*" component={Room} />
   </Router>
 );
